+++ src/gui/main.cpp
@@ -6,7 +6,7 @@ int main(int argc, char **argv)
 {
     QApplication app(argc, argv);
     QApplication::setApplicationName(QStringLiteral("ArenaTES3JSON"));
-    QApplication::setApplicationVersion(QStringLiteral("0.4.0"));
+    QApplication::setApplicationVersion(QStringLiteral("0.4.1"));
     QApplication::setWindowIcon(QIcon(QStringLiteral(":/ArenaTES3JSON.png")));
     arena::tes3json::MainWindow window;
     window.show();